requirements.txt
fastapi==0.88.0
Pillow==9.4.0
psutil==5.9.4
pynvml==11.4.1
tensorflow==2.11.0
tensorflow_intel==2.11.0
